Add support for std::byte to MD5

+++ include/md5.hpp
@@ -217,7 +217,7 @@ namespace md5
         std::uint32_t dataSize = 0;
         for (auto i = begin; i != end; ++i)
         {
-            block[dataSize % blockByteCount] = *i;
+            block[dataSize % blockByteCount] = static_cast<std::uint8_t>(*i);
             if (++dataSize % blockByteCount == 0)
                 transform(block, state);
         }
